新增请假天数少于一天的流程图

+++ src/main/java/com/lv/adminsys/modules/entity/LvLeaveEntity.java
@@ -55,7 +55,7 @@ public class LvLeaveEntity implements Serializable {
     /**
      * 请假天数
      */
-    private Integer lvDay;
+    private float lvDay;
 
     /**
      * 请假截止时间
